requirements.txt
absl-py==2.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.6.1
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
attr==0.3.1
attrs==23.2.0
azure-core==1.32.0
azure-storage-blob==12.25.1
beautifulsoup4==4.12.3
black==25.1.0
bleach==5.0.1
blinker==1.8.2
boto==2.49.0
boto3==1.37.24
botocore==1.37.24
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
cssselect2==0.8.0
datamodel-code-generator==0.26.1
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.7
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.9
django-annoying==0.10.6
django-cors-headers==4.3.1
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-framework==0.1.0
django-restframework==0.0.1
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
docstring_parser==0.16
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
executing==2.0.1
expiringdict==1.2.2
Faker==37.1.0
filelock==3.12.4
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.1
frozenlist==1.4.1
fsspec==2025.3.2
gast==0.6.0
genson==1.3.0
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.83.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.2
google-cloud-logging==3.11.4
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
greenlet==3.0.3
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humansignal-drf-yasg==1.21.10.post1
idna==3.7
ijson==3.3.0
importlib_metadata==8.6.1
inflect==5.6.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.25.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.4
jira==3.8.0
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.9.1
keyring==25.2.1
label-studio==1.16.0
label-studio-sdk==1.0.10
labelImg==1.8.6
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.82
launchdarkly-server-sdk==8.2.1
libclang==18.1.1
lockfile==0.12.2
lxml==5.2.2
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.3
opencv-contrib-python==4.10.0.84
opentelemetry-api==1.31.1
opt_einsum==3.4.0
optree==0.14.1
ordered-set==4.0.2
orjson==3.10.5
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
phonenumbers==8.13.42
Pillow==10.0.1
platformdirs==3.10.0
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==4.25.3
psutil==7.0.0
psycopg==3.1.12
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.1
pypdfium2==4.30.1
pyphen==0.17.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyRFC3339==2.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.4
python-whois==0.9.4
pytz==2022.7.1
pywin32==310
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==27.0.0
redis==5.2.1
referencing==0.36.2
regex==2024.5.15
reportlab==4.0.4
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rq==1.16.2
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.11.4
safetensors==0.5.3
scipy==1.15.2
selenium==4.33.0
semver==3.0.4
sentry-sdk==2.25.0
setuptools==78.1.0
shapely==2.0.7
simplejwt==2.0.1
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.24.0
SQLAlchemy==2.0.31
sqlparse==0.4.3
stack-data==0.6.3
sympy==1.13.1
tenacity==8.4.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.1.3
toml==0.10.2
torch==2.6.0
tornado==6.5.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2022.7
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
user-agents==2.2.0
virtualenv==20.24.5
wcwidth==0.2.13
weasyprint==59.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.40.0
wrapt==1.17.2
wsproto==1.2.0
xmljson==0.2.1
yarl==1.9.4
zipp==3.21.0
zopfli==0.2.3.post1
